Inkling and Kimi-K3 both register a new architecture, so they collide in
the arch/model/mtmd registries and cannot be merged into the nightly as
two independent heads. Both sides are kept:

- src/llama-arch.cpp: KIMI_K3 and INKLING in llm_arch_is_hybrid and
  llm_arch_supports_sm_tensor
- src/llama-model.cpp: both llama_model_* constructors and both
  LLAMA_ROPE_TYPE_NONE cases
- tools/mtmd/CMakeLists.txt: build models/kimik3.cpp and models/inkling.cpp


# Conflicts:
#	src/llama-arch.cpp
#	src/llama-model.cpp
#	tools/mtmd/CMakeLists.txt

+++ ggml/src/ggml-cuda/mmvf.cu
@@ -44,7 +44,7 @@ static __global__ void mul_mat_vec_f(
 
     constexpr int warp_size   = ggml_cuda_get_physical_warp_size();
 
-    x   += int64_t(sample_x)  *stride_sample_x   + channel_x  *stride_channel_x   + row*stride_row;
+    x   += int64_t(sample_x)  *stride_sample_x   + int64_t(channel_x)*stride_channel_x + int64_t(row)*stride_row;
     y   += int64_t(sample_y)  *stride_sample_y   + channel_y  *stride_channel_y;
     dst += int64_t(sample_dst)*stride_sample_dst + channel_dst*stride_channel_dst;
     if constexpr (is_multi_token_id) {
@@ -81,7 +81,7 @@ static __global__ void mul_mat_vec_f(
     }
 
     if (use_gate) {
-        gate_x += int64_t(sample_x)  *stride_sample_x   + channel_x  *stride_channel_x   + row*stride_row;
+        gate_x += int64_t(sample_x)  *stride_sample_x   + int64_t(channel_x)*stride_channel_x + int64_t(row)*stride_row;
     }
 
     if constexpr (has_fusion) {
